% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{as.data.frame.pp}
\alias{as.data.frame.pp}
\title{as.data.frame.pp}
\usage{
\method{as.data.frame}{pp}(pp)
}
\arguments{
\item{pp}{\itemize{
\item pp object
}}
}
\description{
coerce a pp object into a dataframe (loses all but point info)
}
\examples{
pp1 <- mixedcurve::pp(
    points = data.frame(x = runif(100, 0, 1)),
    win = list(c(0, 1)),
    covs = list()
)
df1 <- mixedcurve::as.data.frame(pp)
}
\author{
Jordan Schupbach
}
