% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.r
\name{classify_terms}
\alias{classify_terms}
\title{Classify terms in a formula}
\usage{
classify_terms(form)
}
\arguments{
\item{form}{A formula object.}
}
\value{
A data frame with two columns: "term" and "type". The "term" column
contains the terms from the formula, and the "type" column contains the
corresponding classifications.
}
\description{
This function takes a formula and classifies each term in the formula as
"response", "fixed effect", "random effect", or "kernel fixed effect".
}
\examples{
form <- y ~ K_h(x | a ) + c * d + (K_h(x) | ind / rep) + (x | rep)
classify_terms(form)
}
