% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_dir}
\alias{create_dir}
\title{This function creates a directory if it does not already exist.}
\usage{
create_dir(dir_path)
}
\arguments{
\item{dir_path}{A string representing the path of the directory
to be created.}
}
\value{
None. The function is called for its side effects of
              creating a directory.
}
\description{
This function creates a directory if it does not already exist.
}
\examples{
mixedcurve::create_dir("new_directory")
}
