% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dark_mode}
\alias{dark_mode}
\title{Function to set dark mode theme for R plots}
\usage{
dark_mode()
}
\value{
None. The function is called for its side effects of
              changing graphical parameters.
}
\description{
This function customizes the graphical parameters in R to
create a dark mode theme for plots. It sets the background
color to a dark gray and adjusts text and line colors to be
light for better visibility. It also defines a custom color
palette suitable for dark mode.
}
\examples{
mixedcurve::dark_mode()
plot(1:10, 1:10, main = "Dark Mode Plot", xlab = "X-axis", ylab = "Y-axis")
}
