% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{dist_2d}
\alias{dist_2d}
\title{Function to calculate the Euclidean distance between two points in 2D space}
\usage{
dist_2d(xy1, xy2)
}
\arguments{
\item{xy1}{A numeric vector of length 2 representing the coordinates of the first point (x1, y1)}

\item{xy2}{A numeric vector of length 2 representing the coordinates of the second point (x2, y2)}
}
\value{
The Euclidean distance between the two points
}
\description{
Function to calculate the Euclidean distance between two points in 2D space
}
\examples{
point1 <- c(1, 2)
point2 <- c(4, 6)
distance <- dist_2d(point1, point2)
}
