% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{distance}
\alias{distance}
\title{Function to calculate the Euclidean distance between two vectors}
\usage{
distance(x, y)
}
\arguments{
\item{x}{A numeric vector (or scalar)}

\item{y}{A numeric vector (or scalar) of the same length as x}
}
\description{
Function to calculate the Euclidean distance between two vectors
}
