% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.r
\name{fun}
\alias{fun}
\title{fun}
\usage{
fun(lambda, dom)
}
\arguments{
\item{lambda}{- a function object}

\item{dom}{- domain of the function}
}
\description{
A function class
}
\examples{
myfun <- fun((function(x) sqrt(x)),
    dom = c(0, 1000)
)
plot.fun(myfun)
}
\author{
Jordan Schupbach
}
