% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\name{gauss_kern}
\alias{gauss_kern}
\title{Gaussian kernel function}
\usage{
gauss_kern(x)
}
\arguments{
\item{x}{A numeric value or vector for which to compute the Gaussian kernel.}
}
\value{
The Gaussian kernel value(s) corresponding to the input.
}
\description{
This function computes the Gaussian kernel value for a given input.
}
\examples{
gauss_kern(0) # Should return 1
}
