% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{gen_1d_mc_data_gp}
\alias{gen_1d_mc_data_gp}
\title{n <- ni * nj * nk
list_m <- lapply(1:ngroups, function(i) {
function(t) m(t, i)
})}
\usage{
gen_1d_mc_data_gp(
  m,
  ngroups = 3,
  ni = 6,
  nj = 5,
  nk = 25,
  systematic = TRUE,
  bounds = c(0, 1),
  ind_tau = 1e-06,
  rep_tau = 1e-07,
  noise_tau = 1e-07,
  gp_n = 1000,
  gp_cov_pars = c(1e-04, 0.05)
)
}
\description{
n <- ni * nj * nk
list_m <- lapply(1:ngroups, function(i) {
function(t) m(t, i)
})
}
