% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{gen_fanova_data}
\alias{gen_fanova_data}
\title{Generate functional ANOVA data}
\usage{
gen_fanova_data(
  f = mixedcurve::m3,
  bounds = c(0, 1),
  n = 10,
  ngrp = 3,
  nx = 200,
  balanced = FALSE,
  sigma = 0.05,
  systematic = FALSE,
  px = runif,
  pxargs = list(list(min = 0, max = 1)),
  white_noise = TRUE,
  cov_scale = 0.05,
  gpn = 100,
  family = "gaussian",
  pgrp = NULL,
  pgrpargs = NULL
)
}
\arguments{
\item{f}{A function that takes a numeric matrix of points and a group}

\item{bounds}{A numeric vector of length 2 specifying the domain}

\item{n}{An integer specifying the number of samples to generate.}

\item{ngrp}{An integer specifying the number of groups. Default is 3.}

\item{nx}{An integer specifying the number of points to sample}

\item{balanced}{A logical indicating whether to generate a balanced}

\item{sigma}{A numeric value specifying the standard deviation of}

\item{systematic}{A logical indicating whether to sample points}

\item{px}{A function to sample points for each function. Default is runif.}

\item{pxargs}{A list of arguments to pass to px. Default is}

\item{white_noise}{A logical indicating whether to add white noise}

\item{cov_scale}{A numeric value specifying the scale parameter}

\item{gpn}{An integer specifying the number of grid points
Default is 1000.}

\item{pgrp}{A function to sample group identifiers for an unbalanced}

\item{pgrpargs}{A list of arguments to pass to pgrp. Default is}
}
\value{
A fundata class object (list) containing:
        - df: A data frame with columns for the function values (y),
              group identifiers (grp), sample identifiers (id),
              and input points (x1, x2, ..., xd).
        - coreset: A data frame containing one representative point
                    from each sample.
        - dims: An integer specifying the number of dimensions.
        - nx: An integer specifying the number of points per function.
        - systematic: A logical indicating whether the points
}
\description{
This function generates synthetic functional data for functional ANOVA
analysis. It allows for multiple groups, different sampling schemes,
and the addition of noise.
}
\examples{
# Generate 2D functional ANOVA data on a 20x20 grid
nxy <- 20
data2 <- mixedcurve::gen_fanova_data(
  f = mixedcurve::m2, bounds = list(c(0, 1), c(0, 1)),
  n = 10, ngrp = 3, nx = nxy,
  balanced = FALSE, pgrp = sample,
  pgrpargs = list(x = 1:3, size = 10, replace = TRUE),
  sigma = 0.0005, systematic = TRUE,
  px = runif,
  pxargs = list(
    list(min = 0, max = 1),
    list(min = 0, max = 1)
  ),
  white_noise = TRUE,
  cov_scale = 0.05, gpn = 1000
)
mixedcurve::dark_mode()
par(mfrow = c(3, 3))
mixedcurve::plot.fundata(data2, ncurves = 9)


# Generate 2D functional ANOVA data with random sampling
nxy <- 500
data2 <- mixedcurve::gen_fanova_data(
  f = mixedcurve::m2, bounds = list(c(0, 1), c(0, 1)),
  n = 10, ngrp = 3, nx = nxy,
  balanced = FALSE, pgrp = sample,
  pgrpargs = list(x = 1:3, size = 10, replace = TRUE),
  sigma = 0.0005, systematic = FALSE,
  px = runif,
  pxargs = list(
    list(min = 0, max = 1),
    list(min = 0, max = 1)
  ),
  white_noise = TRUE,
  cov_scale = 0.05, gpn = 1000
)
mixedcurve::dark_mode()
mixedcurve::plot.fundata(data2, ncurves = 9)
tnxy <- nx ^ dims
}
