% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{gen_hfanova_data}
\alias{gen_hfanova_data}
\title{{{{ Sample x
}}} Sample x
{{{ Create group df
}}} Create group df
{{{ Sample noise
}}} Sample noise
{{{ Sample y
mu <- exp(f(as.matrix(xdata), 1) + epsilon)
shape <- 1 / (sigmas\link{1}^2)
scale <- mu / shape
y <- rgamma(length(mu), shape = shape, scale = scale)
}}} Sample y}
\usage{
gen_hfanova_data(
  f,
  n,
  sigmas,
  bounds,
  ndim = 1,
  ngrp = 3,
  px = NULL,
  pxargs = NULL,
  family = "gaussian",
  white_noise = TRUE
)
}
\description{
{{{ Sample x
}}} Sample x
{{{ Create group df
}}} Create group df
{{{ Sample noise
}}} Sample noise
{{{ Sample y
mu <- exp(f(as.matrix(xdata), 1) + epsilon)
shape <- 1 / (sigmas\link{1}^2)
scale <- mu / shape
y <- rgamma(length(mu), shape = shape, scale = scale)
}}} Sample y
}
