% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westfall_young.r
\name{gen_perm_fanova}
\alias{gen_perm_fanova}
\title{Permutation function for functional ANOVA}
\usage{
gen_perm_fanova(dataf, grp_col = "grp", id_col = "id")
}
\arguments{
\item{dataf}{- data frame with columns x, y, grp, id}

\item{grp_col}{- name of the group column}

\item{id_col}{- name of the subject ID column}
}
\value{
A permuted data frame with the same structure as input
}
\description{
Permutation function for functional ANOVA
}
