% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westfall_young.r
\name{gen_pvals_anova}
\alias{gen_pvals_anova}
\title{Generate p-values using ANOVA (requires data to be on a grid)}
\usage{
gen_pvals_anova(dataf, xseq)
}
\arguments{
\item{dataf}{\itemize{
\item data frame with columns x, y, grp, id
}}

\item{xseq}{\itemize{
\item sequence of x values to test
}}
}
\value{
A numeric vector of p-values corresponding to each x in xseq
}
\description{
Generate p-values using ANOVA (requires data to be on a grid)
}
