% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gen_square_layout}
\alias{gen_square_layout}
\title{Generate a square layout matrix for plotting multiple curves}
\usage{
gen_square_layout(ncurves)
}
\arguments{
\item{ncurves}{A positive integer representing the number of curves to be plotted.}
}
\value{
A matrix representing the layout for plotting the curves.
}
\description{
This function generates a layout matrix that arranges a specified number of curves
in a nearly square grid for plotting purposes.
}
\examples{
ncurves <- 10
layout_matrix <- gen_square_layout(ncurves)
layout(layout_matrix)
for (i in 1:ncurves) {
  plot(runif(10), runif(10), main = paste("Curve", i))
}
}
