% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{get_bt_counts}
\alias{get_bt_counts}
\title{get_bt_counts}
\usage{
get_bt_counts(frame, cells = 100)
}
\arguments{
\item{frame}{- a pp object to construct fine-pixel counts}

\item{cells}{- number of cells in fine-pixel approx}
}
\description{
Gets centroids and counts for a fine-pixel approximation
}
\examples{
pp1 <- mixedcurve::pp(data.frame(x = runif(100, 0, 1)),
                      list(c(0, 1)), list())
mixedcurve::get_bt_counts(pp1)
}
\author{
Jordan Schupbach
}
