% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpk.r
\name{glpk}
\alias{glpk}
\title{datas <- data}
\usage{
glpk(
  formula,
  queries,
  data,
  h,
  degree = 0,
  kernel = mixedcurve::gauss_kern,
  family = "gaussian",
  parallel = TRUE,
  cl = NULL
)
}
\description{
datas <- data
}
