% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpk.r
\name{glpkme}
\alias{glpkme}
\title{datas <- data}
\usage{
glpkme(
  formula,
  queries,
  data,
  h,
  degree = 2,
  kernel = mixedcurve::box_kern,
  kthresh = 1e-08,
  family = "gaussian",
  parallel = TRUE,
  alternative_hypothesis = NULL,
  warnings = TRUE,
  cl = NULL
)
}
\description{
datas <- data
}
