% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\name{kern_h}
\alias{kern_h}
\title{Scaled Gaussian kernel function}
\usage{
kern_h(x, h, kernel_fun = gauss_kern)
}
\arguments{
\item{x}{A numeric value or vector for which to compute
the scaled Gaussian kernel.}

\item{h}{A positive numeric value representing the
bandwidth.}
}
\value{
The scaled Gaussian kernel value(s) corresponding
to the input and bandwidth.
}
\description{
This function computes the scaled Gaussian kernel value
for a given input and bandwidth.
}
\examples{
kern_h(0, 1) # Should return 1
kern_h(c(-1, 0, 1), 0.5)
}
