% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{light_mode}
\alias{light_mode}
\title{Function to set light mode theme for R plots}
\usage{
light_mode()
}
\value{
None. The function is called for its side effects of
changing graphical parameters.
}
\description{
This function customizes the graphical parameters in R to
create a dark mode theme for plots. It sets the background
color to white and adjusts text and line colors to be
dark for better visibility. It also defines a custom color
palette suitable for light mode.
}
\examples{
mixedcurve::light_mode()
plot(1:10, 1:10, main = "Light Mode Plot", xlab = "X-axis", ylab = "Y-axis")
}
