% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{list_vignettes}
\alias{list_vignettes}
\title{List Available Vignettes}
\usage{
list_vignettes()
}
\value{
A character vector containing the names of available vignettes.
}
\description{
This function lists the names of available vignettes
in the 'mixedcurve' package by scanning the built_vignettes
directory for HTML files.
}
\examples{
vignettes <- mixedcurve::list_vignettes()
print(vignettes)
}
