% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\name{lm_kernel_weights}
\alias{lm_kernel_weights}
\title{form <- y ~ K_h(x)
bwidth <- 0.1
str(data)
response_term
weighted_design_matrix <- sweep(dmat, 1, weights, FUN = "*")
colnames(weighted_design_matrix) <- paste("w_", colnames(dmat), sep = "")
as.data.frame(weighted_design_matrix)}
\usage{
lm_kernel_weights(
  form,
  data,
  bwidth,
  query,
  kernel_fun = mixedcurve::gauss_kern
)
}
\description{
form <- y ~ K_h(x)
bwidth <- 0.1
str(data)
response_term
weighted_design_matrix <- sweep(dmat, 1, weights, FUN = "*")
colnames(weighted_design_matrix) <- paste("w_", colnames(dmat), sep = "")
as.data.frame(weighted_design_matrix)
}
