% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpk.r
\name{lpk}
\alias{lpk}
\title{datas <- data}
\usage{
lpk(
  formula,
  queries,
  data,
  h,
  degree = 0,
  kernel = mixedcurve::gauss_kern,
  parallel = TRUE,
  alternative_hypothesis = NULL,
  kthresh = 0,
  cl = NULL
)
}
\description{
datas <- data
}
