% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpk.r
\name{lpkme}
\alias{lpkme}
\title{datas <- data}
\usage{
lpkme(
  formula,
  queries,
  data,
  h,
  degree = 0,
  kernel = mixedcurve::gauss_kern,
  parallel = TRUE,
  kthresh = 0,
  cl = NULL
)
}
\description{
datas <- data
}
