% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{m3_1d}
\alias{m3_1d}
\title{Functions m1 to m5 as defined in Cuevas et al. (2004)}
\usage{
m3_1d(t, i)
}
\arguments{
\item{t}{A numeric vector of points in [0, 1]}

\item{i}{An integer in {1, 2, 3} (technically any i works)}
}
\value{
A numeric vector of the same length as t
}
\description{
Functions m1 to m5 as defined in Cuevas et al. (2004)
}
\examples{
x <- seq(0, 1, length.out = 100)
plot(x, mixedcurve::m3(x, 1), type = 'l', xlab ='t', ylab = 'm3(t,i)', main = 'Cuevas: m3(t,i)')
lines(x, mixedcurve::m3(x, 2), col = 2)
lines(x, mixedcurve::m3(x, 3), col = 3)
legend("topright", legend = c("m3(t,1)", "m3(t,2)", "m3(t,3)"), col = 1:3, lty = 1)
}
