% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{m5}
\alias{m5}
\title{Functions m1 to m5 as defined in Cuevas et al. (2004)}
\usage{
m5(t, i)
}
\arguments{
\item{t}{A numeric vector of points in \link{0, 1}}

\item{i}{An integer in {1, 2, 3} (technically any i works)}
}
\value{
A numeric vector of the same length as t
}
\description{
Functions m1 to m5 as defined in Cuevas et al. (2004)
}
\examples{
x <- seq(0, 1, length.out = 100)
plot(x, mixedcurve::m5(x, 1), type = "l", xlab = "t",
     ylab = "m5(t, i)", main = "Cuevas: m5(t, i)", ylim = c(0, 0.1))
lines(x, mixedcurve::m5(x, 2), col = 2)
lines(x, mixedcurve::m5(x, 3), col = 3)
legend("topright", legend = c("m5(t,1)", "m5(t,2)", "m5(t,3)"),
       col = 1:3, lty = 1)
m5_1d <- function(t, i) {
  (i - 1) * 0.01 * dbeta(t, 6, 6)
}
}
