% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{mdoppler}
\alias{mdoppler}
\title{Modified doppler function (generalized to any dim)}
\usage{
mdoppler(x, alpha = 20, beta = 0.25)
}
\arguments{
\item{x}{A numeric matrix of points in [0, 1]^d or a numeric vector in [0, 1]}

\item{alpha}{A numeric value controlling the frequency
of oscillations (default is 20)}

\item{beta}{A numeric value controlling the shift of
the function (default is 0.25)}
}
\value{
A numeric vector of the same length as the number
        of rows in x (if x is a matrix)
or the same length as x (if x is a vector)
}
\description{
A modification of the doppler function (see Donoho and Johnstone 1994)
generalized to any dimension. The generalization involves rotating the 1d
function around the center of the unit cube.
}
\examples{
# 2d example
neval <- 100
image(matrix(mixedcurve::mdoppler(as.matrix(expand.grid(seq(0, 1, length.out = neval),
                               seq(0, 1, length.out = neval)))),
       nrow = neval, ncol = neval))

# 1d example
neval <- 1000
xseq <- seq(0, 1, length.out = neval)
plot(xseq, mixedcurve::mdoppler(xseq), type = 'l', xlab ='x', ylab = 'mdoppler(x)', main = 'Modified Doppler Function (1d)')
}
