% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{open_vignette}
\alias{open_vignette}
\title{Open a Vignette in the Default Web Browser}
\usage{
open_vignette(vname)
}
\arguments{
\item{vname}{A string representing the name of the vignette to be opened.}
}
\value{
None. The function is called for its side effects of
opening a web browser.
}
\description{
This function opens a specified vignette from the 'mixedcurve'
package in the default web browser. The vignette is identified
by its name.
}
\examples{
mixedcurve::open_vignette(list_vignettes()[1])

}
