% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_args}
\alias{parse_args}
\title{Parse Command Line Arguments}
\usage{
parse_args(silent = FALSE)
}
\arguments{
\item{silent}{Logical. If TRUE, suppress messages regarding the absence of arguments.}
}
\value{
A named list containing the parsed arguments.
Each argument will be stored in the list with its name as the key,
and converted to the appropriate type (logical, numeric, or character).
}
\description{
This function parses command line arguments passed to an R script.
It extracts arguments specified in the format `--arg_name value` and
returns them as a named list. The function automatically determines
the type of each argument (logical, numeric, or character).
}
\examples{
# Run the following in the R console with command line arguments:
# Rscript my_script.R --darkmode TRUE --nsim 100
args <- parse_args()
print(args)

}
