% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.r
\name{plot.fun}
\alias{plot.fun}
\title{plot.fun}
\usage{
\method{plot}{fun}(lambda, dom = NULL, evals = 1000, add = FALSE, ...)
}
\arguments{
\item{lambda}{\itemize{
\item a fun object
}}

\item{dom}{\itemize{
\item domain to plot over (defaults to lambda$dom)
}}

\item{evals}{\itemize{
\item number of evaluation points
}}

\item{add}{\itemize{
\item whether to add to existing plot
}}

\item{...}{\itemize{
\item additional arguments to plot()
}}
}
\description{
Plot method for fun objects
}
\examples{
myfun <- mixedcurve::fun((function(x) sqrt(x)),
  dom = c(0, 1000)
)
mixedcurve::plot.fun(myfun)
}
\author{
Jordan Schupbach
}
