% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.r
\name{plot.fundata}
\alias{plot.fundata}
\title{Plot functional ANOVA data}
\usage{
\method{plot}{fundata}(fundata, curves, color_scale = NULL)
}
\arguments{
\item{fundata}{A fundata class object generated by gen_fanova_data().}

\item{curves}{A numeric vector specifying which curves (samples)
to plot.}

\item{color_scale}{An optional color scale for 2D plots.}
}
\value{
Plots of the specified curves.
}
\description{
This function plots functional ANOVA data generated by gen_fanova_data().
It supports both 1D and 2D data, with options for systematic and
random sampling.
}
\examples{
# Generate 1D functional ANOVA data
data1 <- mixedcurve::gen_fanova_data(
  f = mixedcurve::m2, bounds = c(0, 1
  ), n = 5, ngrp = 3, nx = 100,
  balanced = TRUE, sigma = 0.01,
  systematic = TRUE,
  white_noise = TRUE
)
mixedcurve::dark_mode()
par(mfrow = c(2, 3))
mixedcurve::plot.fundata(data1, curves = 1:6)

# Generate 2D functional ANOVA data
nxy <- 30
data2 <- mixedcurve::gen_fanova_data(
  f = mixedcurve::m2, bounds = list(c(0,
  1), c(0, 1)),
  n = 5, ngrp = 3, nx = nxy,
  balanced = TRUE, sigma = 0.0005,
  systematic = TRUE,
  white_noise = TRUE
)
mixedcurve::dark_mode()
par(mfrow = c(2, 3))
mixedcurve::plot.fundata(data2, curves = 1:6)
}
