% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{plot.pp}
\alias{plot.pp}
\title{plot.pp}
\usage{
\method{plot}{pp}(pp, main = "Density of X", tick_col = "black", ...)
}
\arguments{
\item{pp}{\itemize{
\item the pp to plot
}}

\item{main}{\itemize{
\item main title of plot
}}

\item{tick_col}{\itemize{
\item color of tick marks
}}

\item{...}{\itemize{
\item parms passed to base plot()
}}
}
\description{
plot a pp object
}
\examples{
print("Hello World")
}
\author{
Jordan Schupbach
}
