% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{pp}
\alias{pp}
\title{pp}
\usage{
pp(points, win, covs)
}
\arguments{
\item{points}{- a pointset data.frame() object}

\item{win}{- a list of endpoints giving window}

\item{covs}{- covariates associated to the pointset}
}
\description{
A general datastructure for ppps
}
\examples{
pp1 <- mixedcurve::pp(
  points = data.frame(x = runif(100, 0, 1)),
  win = list(c(0, 1)),
  covs = list()
)
plot(pp1, tick_col = "red")

pp2 <- mixedcurve::pp(
  points = data.frame(x = runif(100, 0, 1), y = runif(100, 0, 1)),
  win = list(c(0, 1)),
  covs = list()
)
plot(pp2)
}
\author{
Jordan Schupbach
}
