% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romano_wolf.r
\name{romano_wolf}
\alias{romano_wolf}
\title{Romano-Wolf Stepdown Multiple Testing Procedure}
\usage{
romano_wolf(tstar, boot_tstar)
}
\arguments{
\item{tstar}{A numeric vector of observed test statistics.}

\item{boot_tstar}{A numeric matrix of bootstrap test statistics, where each row
represents a bootstrap sample and each column corresponds to a test statistic.}
}
\value{
A numeric vector of Romano-Wolf adjusted p-values.
}
\description{
This function implements the Romano-Wolf stepdown procedure for adjusting p-values
in multiple hypothesis testing scenarios. It takes observed test statistics and
bootstrap test statistics as inputs and returns adjusted p-values.
}
\note{
: This implementation is nearly identical to the one in the 'wildrwolf'
package except for minor variable renames. This just to avoid dependency on that
package, but this package and that are both GPL3. That implementation
(in 'wildrwolf') is nearly identical to the one in 'hdm' package (which is MIT licensed).
}
\examples{
# Example
set.seed(123)
tstar <- rnorm(5, mean = 2, sd = 1)
boot_tstar <- matrix(rnorm(1000 * 5, mean = 0, sd = 1), nrow = 1000, ncol = 5)
adjusted_pvals <- romano_wolf(tstar, boot_tstar)
print(adjusted_pvals)
}
\references{
https://github.com/s3alfisc/wildrwolf/blob/main/R/get_rwolf_pval.R
https://github.com/cran/hdm/blob/master/R/p_adjust.R
}
