% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{rppp_f}
\alias{rppp_f}
\title{Generate an inhomogeneous poisson point proccess}
\usage{
rppp_f(fun)
}
\arguments{
\item{fun}{- a fun object whose function is the intensity}
}
\description{
Uses method of (TODO: source method for ihppp)
}
\examples{
lmax <- estimate_max(bart_simpson, c(0,6), 10000)
system.time(x <- rppp_f(fun(bart_simpson, c(0,6)))
}
\author{
Jordan Schupbach
}
