% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{rppp_r}
\alias{rppp_r}
\title{Generate an inhomogeneous poisson point proccess}
\usage{
rppp_r(lambda, dom, lmax)
}
\arguments{
\item{dom}{a domain for the intensity function}

\item{lmax}{estimated max for the domain}

\item{fun}{the intensity function}
}
\description{
Uses method of (TODO: source method for ihppp)
}
\examples{
lmax <- mixedcurve::estimate_max(mixedcurve::bart_simpson, c(0,6), 10000)
system.time(x <- mixedcurve::rppp_r(mixedcurve::bart_simpson, c(0,6), lmax))
}
\author{
Jordan Schupbach
}
