% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{seq_centroids}
\alias{seq_centroids}
\title{seq_centroids}
\usage{
seq_centroids(from, to, length.out)
}
\arguments{
\item{from}{- left endpoint}

\item{to}{- right endpoint}

\item{length.out}{- length of returned centroids}
}
\description{
Creates a fixed-width sequence and returns centroids in the sequence
}
\examples{
mixedcurve::seq_centroids(0, 1, 100)
}
\author{
Jordan Schupbach
}
