% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{plot.pp}
\alias{plot.pp}
\title{plot.pp}
\usage{
\method{plot}{pp}(pp, main = "Density of X", tick_col = "black", ...)
}
\arguments{
\item{pp}{- the pp to plot}

\item{main}{- main title of plot}

\item{tick_col}{- color of tick marks}

\item{...}{- parms passed to base plot()}
}
\description{
plot a pp object
}
\examples{
print("Hello World")
}
\author{
Jordan Schupbach
}
