% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{rppp_f}
\alias{rppp_f}
\title{Generate an inhomogeneous poisson point proccess}
\usage{
rppp_f(fun)
}
\arguments{
\item{fun}{\itemize{
\item a fun object whose function is the intensity
}}
}
\description{
Uses method of (TODO: source method for ihppp)
}
\examples{
lmax <- mixedcurve::estimate_max(mixedcurve::bart_simpson, c(0,6), 10000)
x <- mixedcurve::rppp_f(mixedcurve::fun(mixedcurve::bart_simpson, c(0,6))
}
\author{
Jordan Schupbach
}
