% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{seq_centroids}
\alias{seq_centroids}
\title{seq_centroids}
\usage{
seq_centroids(from, to, length.out)
}
\arguments{
\item{from}{\itemize{
\item left endpoint
}}

\item{to}{\itemize{
\item right endpoint
}}

\item{length.out}{\itemize{
\item length of returned centroids
}}
}
\description{
Creates a fixed-width sequence and returns centroids in the sequence
}
\examples{
mixedcurve::seq_centroids(0, 1, 100)
}
\author{
Jordan Schupbach
}
