% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/westfall_young.r
\name{wy_one_step}
\alias{wy_one_step}
\title{Westfall-Young (Cox-Lee) one-step method}
\usage{
wy_one_step(dataf, xseq, nperm, gen_pvals_fun, gen_perm_fun)
}
\arguments{
\item{dataf}{- data frame with columns x, y, grp, id}

\item{xseq}{- sequence of x values to test}

\item{nperm}{- number of permutations}

\item{gen_pvals_fun}{- function to generate p-values, takes dataf and xseq as arguments}

\item{gen_perm_fun}{- function to generate a permutation of the data, takes dataf as argument}
}
\description{
Westfall-Young (Cox-Lee) one-step method
}
